Splitting read command & refactoring
