+++ config/read_params/get_events.json
@@ -1,7 +1,7 @@
 [
   {
-    "from_block": null,
-    "to_block": null,
+    "from_block": {"block_number": null},
+    "to_block": {"block_number": null},
     "address": null,
     "keys": [],
     "continuation_token": null,
